fix(forge): close kotlin :check codegen gap
- bool/present-if accessors + repeat decode + variant peek-byte: emit `!= 0L` for Long-view flag carriers (Uint32/64); switch `also { repeat { it.add } }` to `apply { repeat { add } }` (inner lambda's `it: Int` shadowed the list); cast peek `Byte` to `UByte`
- lookup conformance harness: emit FQN at call site, drop unqualified import; lookup_alarm_code + lookup_severity_default both export `lookupSeverity` (shared `severity` data id) which collided as a top-level import
- u32/u64 conformance unmarshal: u32 widens through Long (Int rejects >2.1B); u64 parses textual content (Long overflows at u64::MAX)

+++ tests/forge/expected/CodecVariantPeekBasic.kt
@@ -55,7 +55,7 @@ data class CodecVariantPeekBasic(
             // next byte for variant tag without advancing — arm body
             // decoder reads it as own header.
             val _peekRaw = cursor.peekSlice(1) ?: return null
-        val _peek: UByte = _peekRaw[0]
+        val _peek: UByte = _peekRaw[0].toUByte()
             // Dispatch on the tag field; each arm decodes its body codec
             // from the cursor. The default arm (when declared) carries
             // the runtime tag value so encode can round-trip it back
